% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{aovSphericityAdjustment}
\alias{aovSphericityAdjustment}
\title{aovSphericityAdjustment}
\usage{
aovSphericityAdjustment(aovObj, type = "GG", adjDF = TRUE)
}
\arguments{
\item{aovObj}{The returned ANOVA object}

\item{type}{"GG" (Greenhouse-Geisser) or "HF" (Huynh-Feldt)}

\item{adjDF}{TRUE/FALSE Should DFs be adjusted?}
}
\value{
list
}
\description{
Adjust ANOVA table with corrections for sphericity (Greenhouse-Geisser or
Huynh-Feldt). Called by default within aovTable.
Note: sphericity corrections require an object with a "Sphericity Corrections" component.
}
\examples{
# Example 1:
# create dataframe with 3(Comp: neutral vs. comp vs. incomp) factors/levels
dat <- createDF(nVP = 20, nTrl = 1,
                design = list("Comp" = c("neutral", "comp", "incomp")))

dat <- addDataDF(dat,
                 RT = list("Comp neutral" = c(510, 150, 100),
                           "Comp comp"    = c(500, 150, 100),
                           "Comp incomp"  = c(520, 150, 100)))
aovRT <- aov(RT ~ Comp + Error(VP/(Comp)), dat)
aovRT <- aovTable(aovRT)
}
