% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleRnwFiles.R
\name{exampleRnwFiles}
\alias{exampleRnwFiles}
\title{exampleRnwFiles(number)}
\usage{
exampleRnwFiles(exampleNumber = 1)
}
\arguments{
\item{exampleNumber}{Example file to open}
}
\value{
Creates a copy of an example *.Rnw file and opens it in the current
working directory. The report be generated via either:
1) Compile PDF option within RStudio
2) Command line:
   Step 1) R CMD Sweave filename.Rnw (creates the *.tex file)
   Step 2) pdflatex filename.tex (creates the *.pdf file)
}
\description{
Open example *.Rnw files.
}
\examples{
library(psychReport)
# Example 1:
exampleRnwFiles(1)
}
