% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{numValueString}
\alias{numValueString}
\title{numValueString}
\usage{
numValueString(value, numDigits = 2, unit = "")
}
\arguments{
\item{value}{number}

\item{numDigits}{number 2 (default)}

\item{unit}{"ms", "mv", "mV", or "\%" or "" (default)}
}
\value{
character
}
\description{
Returns numerical value with requested unit in Latex format with numDigits
number of decimal places and unit symbol.
}
\examples{
# Example 1:
string <- numValueString(100.341, 0, "ms")

# Example 2:
string <- numValueString(2.3412, 2, "mv")

# Example 3:
string <- numValueString(63.9812, 2, "")

}
