% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pValueString}
\alias{pValueString}
\title{pValueString}
\usage{
pValueString(pVal)
}
\arguments{
\item{pVal}{p-value between 0 and 1}
}
\value{
character
}
\description{
Returns Latex formatted string from a p-value required for R/knitr integration.
For example, \emph{p} = 0.11 or \emph{p} < 0.01
Returns values to 3 sig decimal places or < .001
}
\examples{
# Example 1:
pString <- pValueString(0.670)

# Example 2:
pString <- pValueString(0.1234)

# Example 3:
pString <- pValueString("0.03")

}
