% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pValueSummary.R
\name{pValueSummary}
\alias{pValueSummary}
\title{pValueSummary}
\usage{
pValueSummary(pVal, nsmall = 2)
}
\arguments{
\item{pVal}{vector with p-value between 0 and 1}

\item{nsmall}{Number of small digits to round to}
}
\value{
character
}
\description{
Returns p-values summarized using ***, **, *, or exact value
when \emph{p} > .05 (default 2 significant decimal places).
}
\examples{
# Example 1:
psum <- pValueSummary(0.0067)

# Example 2:
psum <- pValueSummary(c(0.6712, 0.1, 0.0001), nsmall = 3)

}
