% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{requiredPackages}
\alias{requiredPackages}
\title{requiredPackages}
\usage{
requiredPackages(
  packages,
  installPackages = FALSE,
  lib = .libPaths()[1],
  repos = "http://cran.us.r-project.org"
)
}
\arguments{
\item{packages}{A list of packages}

\item{installPackages}{TRUE/FALSE Install package if not installed}

\item{lib}{character vector giving the library directories where to install the packages. Recycled as needed. If missing, defaults to the first element of .libPaths()}

\item{repos}{character vector, the base URL(s) of the repositories to use, e.g., the URL of a CRAN mirror such as "https://cloud.r-project.org". For more details on supported URL schemes see url. Can be NULL to install from local files, directories or URLs: this will be inferred by extension from pkgs if of length one.}
}
\description{
Installs (default if required) and loads specified packages.
}
