% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aovDispMeans.R
\name{aovDispMeans}
\alias{aovDispMeans}
\title{aovDispMeans}
\usage{
aovDispMeans(ezObj, value = "value", caption = sys.call())
}
\arguments{
\item{ezObj}{Output from ezANOVA  (NB. ezANOVA must be called with \"return_aov = TRUE\"")}

\item{value}{String for column name}

\item{caption}{Required for heading}
}
\description{
Displays marginal means from model.tables in the command window.
}
\examples{
requiredPackages(c("ez"))

# Example 1:
# create dataframe
dat <- createDF(nVP = 50,
                nTrl = 1,
                design = list("Comp" = c("comp", "incomp")))

dat <- addDataDF(dat, RT = list("Comp_comp"   = c(500, 100, 100),
                                "Comp_incomp" = c(520, 100, 100)))

dat$VP <- as.factor(dat$VP)
aovRT <- ezANOVA(dat, dv=.(RT), wid = .(VP), within = .(Comp),
                 return_aov = TRUE, detailed = TRUE)
aovRT <- aovTable(aovRT)
aovDispMeans(aovRT)

}
