% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aovDispTable.R
\name{aovDispTable}
\alias{aovDispTable}
\title{aovDispTable}
\usage{
aovDispTable(ezObj, caption = sys.call())
}
\arguments{
\item{ezObj}{Output from ezANOVA}

\item{caption}{Required for heading}
}
\description{
Display formatted ANOVA table in command window.
}
\examples{
# Example 1:
requiredPackages(c("ez"))
# create dataframe
dat <- createDF(nVP = 6,
                nTrl = 1,
                design = list("Comp" = c("comp", "incomp")))

dat <- addDataDF(dat, RT = list("Comp_comp"   = c(500, 150, 100),
                                "Comp_incomp" = c(520, 150, 100)))

dat$VP <- as.factor(dat$VP)
aovRT <- ezANOVA(dat, dv=.(RT), wid = .(VP), within = .(Comp), return_aov = TRUE, detailed = TRUE)
aovDispTable(aovRT)

}
