% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aovEffectSize.R
\name{aovEffectSize}
\alias{aovEffectSize}
\title{aovEffectSize}
\usage{
aovEffectSize(ezObj, effectSize)
}
\arguments{
\item{ezObj}{Output from ezANOVA}

\item{effectSize}{"ges" vs. pes"}
}
\value{
list
}
\description{
Add partial eta squared (pes) or eta-squared (es) effect size measures
to ezANOVA table.
}
\examples{
requiredPackages(c("ez", "dplyr"))

# Example 1:
# create dataframe with 2(Comp: comp vs. incomp) and 2(Side: left vs. right) factors/levels
dat <- createDF(nVP = 20,
                nTrl = 50,
                design = list("Comp" = c("comp", "incomp", "neutral"),
                              "Side" = c("left", "right")))

dat <- addDataDF(dat,
                 RT = list("Comp:Side_comp:left"     = c(500, 150, 150),
                           "Comp:Side_comp:right"    = c(500, 150, 150),
                           "Comp:Side_incomp:left"   = c(550, 150, 150),
                           "Comp:Side_incomp:right"  = c(550, 150, 150),
                           "Comp:Side_neutral:left"  = c(525, 150, 150),
                           "Comp:Side_neutral:right" = c(525, 150, 150)))

# aggregate dat across trials
datAggVP <- dat \%>\%
    group_by(VP, Comp, Side) \%>\%
    summarize(N  = n(),
              rt = mean(RT))

# repeated measures ANOVA using ezANOVA
datAggVP$VP <- as.factor(datAggVP$VP)
aovRT <- ezANOVA(datAggVP, dv=.(rt), wid = .(VP), within = .(Comp, Side), 
                 return_aov = TRUE, detailed = TRUE)
aovRT <- aovEffectSize(aovRT, "pes")
aovRT <- aovTable(aovRT)

}
