% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{effectsizeValueString}
\alias{effectsizeValueString}
\title{effectsizeValueString}
\usage{
effectsizeValueString(aovObj, effect, effectSize = "pes")
}
\arguments{
\item{aovObj}{Output from aov or ezANOVA (NB. ezANOVA must be called with detailed = TRUE)}

\item{effect}{The effect within the ANOVA table to return}

\item{effectSize}{The effect size to report ("ges" default within ezANOVA, "pes" partial eta squared)}
}
\value{
character
}
\description{
Returns required Latex formatted string for effect size (eta squared, partial
eta square, generalized eta squared) = XXX for R/Sweave integration.
Returns values to 2 sig decimal places.
}
\examples{
requiredPackages(c("dplyr", "ez"))
# Example 1:
# create dataframe and add data with 2(Comp: comp vs. incomp) and 2(Side: left vs. right)
dat <- createDF(nVP = 20,
                nTrl = 50,
                design = list("Comp" = c("comp", "incomp"),
                              "Side" = c("left", "right")))

dat <- addDataDF(dat, RT = list("Comp:Side_comp:left"    = c(500, 150, 100),
                                "Comp:Side_comp:right"   = c(500, 150, 100),
                                "Comp:Side_incomp:left"  = c(520, 150, 100),
                                "Comp:Side_incomp:right" = c(520, 150, 100)))

# aggregate dat across trials
datAggVP <- dat \%>\%
    group_by(VP, Comp, Side) \%>\%
    summarize(N  = n(),
              rt = mean(RT))

# repeated measures ANOVA using ezANOVA
datAggVP$VP <- as.factor(datAggVP$VP)
aovRT <- ezANOVA(datAggVP, dv=.(rt), wid = .(VP), within = .(Comp, Side),
                 return_aov = TRUE, detailed = TRUE)
aovRT <- aovTable(aovRT)

pesString <- effectsizeValueString(aovRT, "Comp")  # partial eta squared
pesString <- effectsizeValueString(aovRT, "Comp:Side")

aovRT <- aov(rt ~ Comp*Side + Error(VP/(Comp*Side)), datAggVP)
aovRT <- aovTable(aovRT)

pesString <- effectsizeValueString(aovRT, "Comp")  # partial eta squared
pesString <- effectsizeValueString(aovRT, "Comp:Side")

}
