% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{fValueString}
\alias{fValueString}
\title{fValueString}
\usage{
fValueString(ezObj, effect)
}
\arguments{
\item{ezObj}{The returned object from a call to ezANOVA}

\item{effect}{The effect within the ANOVA table to return}
}
\description{
Returns required Latex formatted string for \emph{F}(df1, df2) = XXX
for R/Sweave integration. For example, \emph{F}(1, 23) = 3.45.
Returns values to 2 sig decimal places.
}
