% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{printTable}
\alias{printTable}
\title{printTable}
\usage{
printTable(obj, caption = "DF", digits = 3, onlyContents = FALSE)
}
\arguments{
\item{obj}{Dataframe or ANOVA table to print}

\item{caption}{Title of the dataframe}

\item{digits}{Number of digits to round to NB. length can be 1, or vector with
length equal to the number of numeric columns}

\item{onlyContents}{TRUE/FALSE}
}
\value{
character
}
\description{
Returns Latex formatted table from dataframe or ANOVA table.
Uses xtable latex package with some basic defaults.
For more examples, see R package xtable
}
\examples{
# Example 1:
# create dataframe
dat <- createDF(nVP = 6, nTrl = 1,
                design = list("Comp" = c("comp", "incomp", "neutral")))

dat <- addDataDF(dat, RT = list("Comp comp"    = c(500, 150, 100),
                                "Comp incomp"  = c(520, 150, 100),
                                "Comp neutral" = c(510, 150, 100)))
printTable(dat, digits = c(0, 2)) # latex formatted
printTable(dat, digits = 0)       # latex formatted

aovRT <- aov(RT ~ Comp + Error(VP/(Comp)), dat)
aovRT <- aovTable(aovRT)
printTable(aovRT$ANOVA) # latex formatted
printTable(aovRT$ANOVA, digits = c(0,2,2,2)) # latex formatted


}
