% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychReport.R
\docType{package}
\name{psychReport-package}
\alias{psychReport}
\alias{psychReport-package}
\title{psychReport}
\description{
Helper functions for producing reports in Psychology (Reproducible Research). Provides required formatted strings (APA style) for use in 'Knitr'/'Latex' integration within *.Rnw files.
}
