% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pValueString.R
\name{pValueString}
\alias{pValueString}
\title{pValueString}
\usage{
pValueString(pVal, nsmall = 2)
}
\arguments{
\item{pVal}{p-value between 0 and 1}

\item{nsmall}{Number of small digits to round to}
}
\value{
character
}
\description{
Returns Latex formatted string from a p-value required for R/Sweave integration.
For example, \emph{p} = 0.11 or \emph{p} < 0.01
Returns values to 2 sig decimal places if p-value >= 0.05.
}
\examples{
# Example 1:
pString <- pValueString(0.67)

# Example 2:
pString <- pValueString(0.1234, 3)

# Example 3:
pString <- pValueString("0.03")

\dontrun{
# Example use in *.Rnw Sweave file
# \Sexpr{pString} }

}
