% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pValueSummary}
\alias{pValueSummary}
\title{pValueSummary}
\usage{
pValueSummary(pVal)
}
\arguments{
\item{pVal}{vector with p-value between 0 and 1}
}
\value{
character
}
\description{
Returns p-values summarized using ***, **, *, or exact value
when \emph{p} > .05 (default 2 significant decimal places).
}
\examples{
# Examples:
psum <- pValueSummary(0.0067)
psum <- pValueSummary(c(0.0001, 0.002, 0.02, 0.1))

}
